refactor(eve): centralize HITL request interpretation
